requirements.txt
mysql-connector-python==8.3.0
ttkbootstrap==1.10.0
ttkthemes==3.2.2
